fix lost model reference
