+++ benchmarks/test_trainer_parity.py
@@ -8,6 +8,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from torch.utils.data import DataLoader
 from torchvision import transforms
+import tests.base.utils as tutils
 
 from pytorch_lightning import Trainer, LightningModule
 from tests.base.datasets import TestingMNIST
@@ -64,20 +65,7 @@ def test_pytorch_parity(tmpdir):
     for pl_out, pt_out in zip(lightning_outs, manual_outs):
         np.testing.assert_almost_equal(pl_out, pt_out, 5)
 
-    assert_speed_parity(pl_times, pt_times, num_epochs)
-
-
-def assert_speed_parity(pl_times, pt_times, num_epochs):
-
-    # assert speeds
-    max_diff_per_epoch = 0.9
-    pl_times = np.asarray(pl_times)
-    pt_times = np.asarray(pt_times)
-    diffs = pl_times - pt_times
-    diffs = diffs / num_epochs
-
-    assert np.alltrue(diffs < max_diff_per_epoch), \
-        f"lightning was slower than PT (threshold {max_diff_per_epoch})"
+    tutils.assert_speed_parity(pl_times, pt_times, num_epochs)
 
 
 def set_seed(seed):
